test(audit): move lifecycle logic to unit, response-id cases to the shared suite
Reshape audit.lifecycle (20 typed-spec tests): 12 pure-logic cases (targets,
metadata, redaction, multi-target) become unit tests; the 6 targetFromResponse
cases (created-resource id from the real response) fold into audit.integration.ts;
the 2 denied cases are dropped (covered by the shared deny probes). Delete the
per-endpoint lifecycle integration file.
