+++ packages/server/lib/middleware/audit.integration.test.ts
@@ -3,10 +3,12 @@ import { afterAll, beforeAll, beforeEach, describe, expect, it, vi } from 'vites
 import db from '@nangohq/database';
 import * as featureFlags from '@nangohq/feature-flags';
 import { customerKeyService, seeders, updatePlan, userService } from '@nangohq/shared';
+import { getLogger } from '@nangohq/utils';
 
 import { audit } from '../audit.js';
 import { authenticateUser, isSuccess, runServer } from '../utils/tests.js';
 
+import type { AuditAction, AuditResource } from '@nangohq/audit';
 import type { MockInstance } from 'vitest';
 
 // The single audit integration suite: only the cases that genuinely need the live stack. Everything
@@ -23,6 +25,12 @@ import type { MockInstance } from 'vitest';
 
 let api: Awaited<ReturnType<typeof runServer>>;
 let auditSpy: MockInstance<typeof audit.record>;
+
+// authenticateUser() signs in, which records an app_auth/login event, so the event under test is not
+// necessarily calls[0]. Select it by resource/action instead of by position.
+function auditEvent(resource: AuditResource, action: AuditAction) {
+    return auditSpy.mock.calls.map((call) => call[0]).find((event) => event.resource === resource && event.action === action);
+}
 
 describe('audit middleware — live-stack contract', () => {
     beforeAll(async () => {
@@ -186,5 +194,201 @@ describe('audit middleware — live-stack contract', () => {
             targets: [{ type: 'member', id: String(other.user.id) }]
         });
         expect(event?.targets[0]).not.toHaveProperty('display');
+    });
+
+    // The id of a created resource is known only from the response body, so these specs resolve their
+    // target with targetFromResponse — which only a real controller producing a real response can drive.
+    describe('created-resource id from the real response', () => {
+        it('records an integration creation with the created key and provider', async () => {
+            const { apiKey } = await seeders.seedAccountEnvAndUser();
+
+            const res = await api.fetch('/integrations', {
+                method: 'POST',
+                token: apiKey.secret,
+                body: { provider: 'algolia', unique_key: 'audit-algolia' }
+            });
+
+            expect(res.res.status).toBe(200);
+            isSuccess(res.json);
+            await vi.waitFor(() => {
+                expect(auditEvent('integration', 'created')).toBeDefined();
+            });
+            expect(auditEvent('integration', 'created')).toMatchObject({
+                resource: 'integration',
+                action: 'created',
+                outcome: 'success',
+                targets: [{ type: 'integration', id: 'audit-algolia' }],
+                metadata: { provider: 'algolia' }
+            });
+        });
+
+        it('records an environment creation with the response id and is account-scoped', async () => {
+            const { plan, apiKey } = await seeders.seedAccountEnvAndUser();
+            await updatePlan(db.knex, { id: plan.id, environments_max: 10 });
+
+            const res = await api.fetch('/api/v1/environments', {
+                method: 'POST',
+                token: apiKey.secret,
+                body: { name: 'staging' }
+            });
+
+            expect(res.res.status).toBe(200);
+            isSuccess(res.json);
+            const createdId = String(res.json.data.id);
+            await vi.waitFor(() => {
+                expect(auditEvent('environment', 'created')).toBeDefined();
+            });
+            expect(auditEvent('environment', 'created')).toMatchObject({
+                resource: 'environment',
+                action: 'created',
+                outcome: 'success',
+                // Account-scoped events null the environment field.
+                environment: null,
+                targets: [{ type: 'environment', id: createdId, display: 'staging' }],
+                metadata: { name: 'staging' }
+            });
+        });
+
+        it('records an api key creation without ever recording the secret value', async () => {
+            const { apiKey } = await seeders.seedAccountEnvAndUser();
+
+            const res = await api.fetch('/api/v1/environment/api-keys', {
+                method: 'POST',
+                token: apiKey.secret,
+                // @ts-expect-error querystring is not typed on this endpoint
+                query: { env: 'dev' },
+                body: { display_name: 'ci-key', scopes: ['environment:*'] }
+            });
+
+            expect(res.res.status).toBe(200);
+            isSuccess(res.json);
+            const createdId = String(res.json.data.id);
+            const secret = res.json.data.secret;
+            await vi.waitFor(() => {
+                expect(auditEvent('api_key', 'created')).toBeDefined();
+            });
+            const event = auditEvent('api_key', 'created');
+            expect(event).toMatchObject({
+                resource: 'api_key',
+                action: 'created',
+                outcome: 'success',
+                targets: [{ type: 'api_key', id: createdId, display: 'ci-key' }],
+                metadata: { displayName: 'ci-key', scopes: ['environment:*'] }
+            });
+            // The secret is present in the HTTP response but must never reach the audit record.
+            expect(secret.length).toBeGreaterThan(0);
+            expect(JSON.stringify(event)).not.toContain(secret);
+        });
+
+        it('records a connection import with the server-generated connection_id when none is supplied', async () => {
+            const { env, apiKey } = await seeders.seedAccountEnvAndUser();
+            await seeders.createConfigSeed(env, 'github', 'github');
+
+            const res = await api.fetch('/connections', {
+                method: 'POST',
+                token: apiKey.secret,
+                body: { provider_config_key: 'github', credentials: { type: 'OAUTH2', access_token: '123' } }
+            });
+
+            expect(res.res.status).toBe(201);
+            isSuccess(res.json);
+            // The request omitted connection_id, so the target can only come from the response body.
+            const generatedId = res.json.connection_id;
+            expect(generatedId).toBeTruthy();
+
+            await vi.waitFor(() => {
+                expect(auditEvent('connection', 'created')).toBeDefined();
+            });
+            expect(auditEvent('connection', 'created')).toMatchObject({
+                resource: 'connection',
+                action: 'created',
+                outcome: 'success',
+                targets: [{ type: 'connection', id: generatedId }],
+                metadata: { providerConfigKey: 'github' }
+            });
+        });
+
+        it('records a private integration creation with the response unique_key as target', async () => {
+            const { apiKey } = await seeders.seedAccountEnvAndUser();
+
+            const res = await api.fetch('/api/v1/integrations', {
+                method: 'POST',
+                token: apiKey.secret,
+                query: { env: 'dev' },
+                body: {
+                    provider: 'algolia',
+                    useSharedCredentials: false,
+                    integrationId: 'audit-private-algolia'
+                }
+            });
+
+            expect(res.res.status).toBe(200);
+            isSuccess(res.json);
+            // The private path omits the key from the request — the target is resolved from the response body.
+            const uniqueKey = res.json.data.unique_key;
+            expect(uniqueKey.length).toBeGreaterThan(0);
+            await vi.waitFor(() => {
+                expect(auditEvent('integration', 'created')).toBeDefined();
+            });
+            expect(auditEvent('integration', 'created')).toMatchObject({
+                resource: 'integration',
+                action: 'created',
+                outcome: 'success',
+                targets: [{ type: 'integration', id: uniqueKey }],
+                metadata: { provider: 'algolia' }
+            });
+        });
+
+        it('records a failed integration creation targetless without reading the response body', async () => {
+            const { apiKey } = await seeders.seedAccountEnvAndUser();
+
+            const created = await api.fetch('/api/v1/integrations', {
+                method: 'POST',
+                token: apiKey.secret,
+                query: { env: 'dev' },
+                body: { provider: 'algolia', useSharedCredentials: false, integrationId: 'audit-dup-algolia' }
+            });
+            expect(created.res.status).toBe(200);
+            isSuccess(created.json);
+            auditSpy.mockClear();
+
+            // The created-resource resolver reads `response.data.unique_key`; a failed create's body is
+            // `{ error }` with no `data`, so running it on a non-success outcome would throw and log an error.
+            // `logger.error` is inherited from a prototype shared by every audit logger, so spying it here
+            // catches the middleware's own logger too.
+            const auditLogger = getLogger('Audit');
+            let errorProto: object = auditLogger;
+            while (errorProto && !Object.prototype.hasOwnProperty.call(errorProto, 'error')) {
+                errorProto = Object.getPrototypeOf(errorProto) as object;
+            }
+            const errorSpy = vi.spyOn(errorProto as { error: (...args: unknown[]) => unknown }, 'error');
+
+            try {
+                // Re-posting the same integrationId is rejected with a 400 whose body carries no `data`.
+                const res = await api.fetch('/api/v1/integrations', {
+                    method: 'POST',
+                    token: apiKey.secret,
+                    query: { env: 'dev' },
+                    body: { provider: 'algolia', useSharedCredentials: false, integrationId: 'audit-dup-algolia' }
+                });
+
+                expect(res.res.status).toBe(400);
+                await vi.waitFor(() => {
+                    expect(auditSpy).toHaveBeenCalled();
+                });
+                // A failed create emits targetless (the id only exists on success) with a failure outcome.
+                expect(auditSpy).toHaveBeenCalledTimes(1);
+                expect(auditEvent('integration', 'created')).toMatchObject({
+                    resource: 'integration',
+                    action: 'created',
+                    outcome: 'failure',
+                    targets: []
+                });
+                // The resolver must be gated on success — it never ran, so nothing was logged for it.
+                expect(errorSpy.mock.calls.some((call) => String(call[0]).includes('failed to resolve audit target from response'))).toBe(false);
+            } finally {
+                errorSpy.mockRestore();
+            }
+        });
     });
 });
